requirements.txt
PyQt5
pyqt5-tools
ShopifyAPI
